requirements.txt
apikey
google-search-results
yahooquery
google-generativeai